Add partial.ImageByPlatform helper
This is used by ko now in a couple places, and it seems like a useful
helper to provide in general.

Reuse this helper in pkg/v1/remote/
